+++ src/init.ts
@@ -48,5 +48,3 @@ function equals<T>(obj: Object | Array<T>): boolean {
 Array.prototype.each = arrayEach;
 Array.prototype.equals = equals;
 String.prototype.capitalize = capitalize;
-Object.prototype.isEmpty = isEmpty;
-Object.prototype.each = objectEach;